+++ app/screens/settings/advanced/index.tsx
@@ -3,31 +3,22 @@
 
 import React, {useEffect, useState} from 'react';
 import {useIntl} from 'react-intl';
-import {View, TouchableOpacity} from 'react-native';
-import {Edge, SafeAreaView} from 'react-native-safe-area-context';
+import {TouchableOpacity} from 'react-native';
 
 import {useTheme} from '@context/theme';
 import useAndroidHardwareBackHandler from '@hooks/android_back_handler';
 import {popTopScreen} from '@screens/navigation';
 import {deleteFileCache, getAllFilesInCachesDirectory, getFormattedFileSize} from '@utils/file';
 import {preventDoubleTap} from '@utils/tap';
-import {changeOpacity, makeStyleSheetFromTheme} from '@utils/theme';
+import {makeStyleSheetFromTheme} from '@utils/theme';
 
+import SettingContainer from '../setting_container';
 import SettingOption from '../setting_option';
 
 import type {ReadDirItem} from 'react-native-fs';
 
 const getStyleSheet = makeStyleSheetFromTheme((theme) => {
     return {
-        screen: {
-            flex: 1,
-            backgroundColor: theme.centerChannelBg,
-        },
-        body: {
-            backgroundColor: changeOpacity(theme.centerChannelColor, 0.06),
-            flex: 1,
-            paddingTop: 35,
-        },
         itemStyle: {
             backgroundColor: theme.centerChannelBg,
             paddingHorizontal: 20,
@@ -37,7 +28,6 @@ const getStyleSheet = makeStyleSheetFromTheme((theme) => {
 
 const EMPTY_FILES: ReadDirItem[] = [];
 const EMPTY_SERVERS: string[] = [];
-const EDGES: Edge[] = ['left', 'right'];
 
 type AdvancedSettingsProps = {
     componentId: string;
@@ -82,29 +72,22 @@ const AdvancedSettings = ({componentId}: AdvancedSettingsProps) => {
     const disabled = Boolean(dataSize && (dataSize > 0));
 
     return (
-        <SafeAreaView
-            edges={EDGES}
-            style={styles.screen}
-            testID='settings_display.screen'
-        >
-            <View
-                style={styles.body}
+        <SettingContainer>
+            <TouchableOpacity
+                onPress={onPressDeleteData}
+                disabled={disabled}
+                activeOpacity={disabled ? 0 : 1}
             >
-                <TouchableOpacity
-                    onPress={onPressDeleteData}
-                    disabled={disabled}
-                    activeOpacity={disabled ? 0 : 1}
-                >
-                    <SettingOption
-                        containerStyle={styles.itemStyle}
-                        destructive={true}
-                        label={intl.formatMessage({id: 'advanced_settings.delete_data', defaultMessage: 'Delete Documents & Data'})}
-                        info={getFormattedFileSize(dataSize || 0)}
-                        type='none'
-                    />
-                </TouchableOpacity>
-            </View>
-        </SafeAreaView>
+                <SettingOption
+                    containerStyle={styles.itemStyle}
+                    destructive={true}
+                    icon='trash-can-outline'
+                    info={getFormattedFileSize(dataSize || 0)}
+                    label={intl.formatMessage({id: 'advanced_settings.delete_data', defaultMessage: 'Delete Documents & Data'})}
+                    type='none'
+                />
+            </TouchableOpacity>
+        </SettingContainer>
     );
 };
