+++ app/screens/settings/notification_mention/notification_mention.tsx
@@ -2,39 +2,13 @@
 // See LICENSE.txt for license information.
 
 import React from 'react';
-import {ScrollView} from 'react-native';
-import {Edge, SafeAreaView} from 'react-native-safe-area-context';
 
-import {useTheme} from '@context/theme';
-import {makeStyleSheetFromTheme} from '@utils/theme';
-import {typography} from '@utils/typography';
+import SettingContainer from '../setting_container';
 
 import MentionSettings from './mention_settings';
 import ReplySettings from './reply_settings';
 
 import type UserModel from '@typings/database/models/servers/user';
-
-const getStyleSheet = makeStyleSheetFromTheme((theme) => {
-    return {
-        container: {
-            flex: 1,
-            backgroundColor: theme.centerChannelBg,
-        },
-        input: {
-            color: theme.centerChannelColor,
-            height: 40,
-            ...typography('Body', 75, 'Regular'),
-        },
-        scrollView: {
-            flex: 1,
-        },
-        scrollViewContent: {
-            paddingVertical: 35,
-        },
-    };
-});
-
-const edges: Edge[] = ['left', 'right'];
 
 type NotificationMentionProps = {
     componentId: string;
@@ -42,27 +16,14 @@ type NotificationMentionProps = {
     isCRTEnabled: boolean;
 }
 const NotificationMention = ({componentId, currentUser, isCRTEnabled}: NotificationMentionProps) => {
-    const theme = useTheme();
-    const styles = getStyleSheet(theme);
-
     return (
-        <SafeAreaView
-            edges={edges}
-            testID='notification_mention.screen'
-            style={styles.container}
-        >
-            <ScrollView
-                style={styles.scrollView}
-                contentContainerStyle={styles.scrollViewContent}
-                alwaysBounceVertical={false}
-            >
-                <MentionSettings
-                    currentUser={currentUser}
-                    componentId={componentId}
-                />
-                {!isCRTEnabled && <ReplySettings/>}
-            </ScrollView>
-        </SafeAreaView>
+        <SettingContainer>
+            <MentionSettings
+                currentUser={currentUser}
+                componentId={componentId}
+            />
+            {!isCRTEnabled && <ReplySettings/>}
+        </SettingContainer>
     );
 };
